refactor(db): final round of Claude Auto Review cleanups for #834
- cards-safe-columns.ts を card-padding-color-errors.ts へ改名
  (CARDS_SAFE_COLUMNS 等の撤去後、残る責務が image_padding_color 専用の
  デプロイ窓ヘルパーのみになったため、card-number-errors.ts 等と同じ粒度の
  ファイル名に揃える。改名に伴い11ファイルのimport文とプローズ内の
  ファイル名参照を追従)。
- dashboard-data.ts / recalculate-drop-rates.ts: フォールバック撤去により
  1回しか呼ばれなくなった内部クロージャ(selectCards/selectRows/
  selectRecentGachaHistory/selectActiveCardsForStreamer、計6箇所)を
  呼び出し元へインライン化しネストを削減。
- card-padding-color-errors.test.ts: 列数のハードコード(22)を
  getTableColumns(cardsTable).length - 1 の動的算出に変更。
- ソースコードコメントから「Claude Auto Review」というレビューツール名への
  言及を除去し、挙動・issue番号ベースの説明に統一。
- card_number系(#393/#548)・image_padding_color系(#899)の残存フォールバック
  の要否実測・整理をフォローアップissue #954へ切り出し、関連コメントから参照。

検証: test:unit 3455/3455 pass, typecheck 0 errors, eslint 0 errors, build成功
