+++ src/app/api/cards/batch-update/route.ts
@@ -13,7 +13,7 @@ import { getDb } from "@/lib/db/client";
 
 import { withDbRetry } from "@/lib/db/retry";
 import { cards as cardsTable, streamers as streamersTable } from "@/lib/db/schema";
-import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR, isMissingCardPaddingColorError } from "@/lib/db/cards-safe-columns";
+import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR, isMissingCardPaddingColorError } from "@/lib/db/card-padding-color-errors";
 import type { ApiRateLimitResponse } from "@/types/api";
 
 /**
@@ -100,24 +100,23 @@ async function fetchExistingCardsForBatchUpdatePg(
  * #834: 「本番未デプロイ8列」（card_number/hp/atk/def/spd/skill_*）を理由にした
  * 列制限は、本番実測で8列とも実在することを確認したため撤去した。
  *
- * Claude Auto Review 指摘対応: この関数のレスポンスはクライアントの
- * handleBatchDropRateSave（CardManager.tsx）でカードオブジェクトごと置換され、
- * SortedCardGrid.tsx 等の描画で image_padding_color を使うため、この列を
- * 返さないと保存直後の画面上で余白色が既定に戻って見える（再読み込みで復旧）。
- * そこで cards/route.ts の insertCardPg 等と同じ「まず無指定 select を試し、
- * image_padding_color 列欠落（#899、本Issueとは独立した別デプロイ窓）を検知した
- * 場合のみ CARDS_COLUMNS_WITHOUT_PADDING_COLOR へ切り替えて再試行する」方針に
- * 揃える。列が存在する環境（本番・preview とも実測済み）では無指定 select が
- * そのまま成功するため、余白色も含めて返る。
+ * この関数のレスポンスはクライアントの handleBatchDropRateSave（CardManager.tsx）
+ * でカードオブジェクトごと置換され、SortedCardGrid.tsx 等の描画で
+ * image_padding_color を使うため、この列を返さないと保存直後の画面上で余白色が
+ * 既定に戻って見える（再読み込みで復旧）。そこで cards/route.ts の insertCardPg
+ * 等と同じ「まず無指定 select を試し、image_padding_color 列欠落（#899、本Issue
+ * とは独立した別デプロイ窓）を検知した場合のみ CARDS_COLUMNS_WITHOUT_PADDING_COLOR
+ * へ切り替えて再試行する」方針に揃える。列が存在する環境（本番・preview とも
+ * 実測済み）では無指定 select がそのまま成功するため、余白色も含めて返る。
  *
- * 意図した応答契約の変更（Claude Auto Review 再指摘）: この関数は変更前、
- * 常時 CARDS_SAFE_COLUMNS（image_padding_color に加え card_number/hp/atk 等
- * バトル7列も除外）を使っていたため、POST /api/cards/batch-update のレスポンス
- * cards には card_number 等も常に欠落していた。無指定 select への変更により、
- * これらの列も image_padding_color と同様にレスポンスへ含まれるようになる
- * （副次効果として、確率一括保存直後のローカル state から card_number が消える
- * 問題も解消する）。呼び出し元（recalculateIfAutoMode 経由の recalculatedCards、
- * 既に #834 で無指定 select 化済み）とも列構成が揃う。
+ * 意図した応答契約の変更: この関数は変更前、常時 CARDS_SAFE_COLUMNS
+ * （image_padding_color に加え card_number/hp/atk 等バトル7列も除外）を
+ * 使っていたため、POST /api/cards/batch-update のレスポンス cards には
+ * card_number 等も常に欠落していた。無指定 select への変更により、これらの列も
+ * image_padding_color と同様にレスポンスへ含まれるようになる（副次効果として、
+ * 確率一括保存直後のローカル state から card_number が消える問題も解消する）。
+ * 呼び出し元（recalculateIfAutoMode 経由の recalculatedCards、既に #834 で
+ * 無指定 select 化済み）とも列構成が揃う。
  */
 async function fetchUpdatedCardsForBatchUpdatePg(
   streamerId: string,
