+++ src/app/api/cards/route.ts
@@ -34,7 +34,7 @@ import { getDb } from "@/lib/db/client";
 
 import { withDbRetry } from "@/lib/db/retry";
 import { cards as cardsTable, streamers as streamersTable, userCards as userCardsTable } from "@/lib/db/schema";
-import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR, isMissingCardPaddingColorError } from "@/lib/db/cards-safe-columns";
+import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR, isMissingCardPaddingColorError } from "@/lib/db/card-padding-color-errors";
 import type { ApiRateLimitResponse } from "@/types/api";
 
 // pg (postgres.js) が throw するエラーの汎用形状。card-number-errors.ts /
@@ -154,14 +154,14 @@ async function insertCardPg(
 
   let { card, error } = await attemptInsert();
 
-  // Claude Auto Review 指摘: card_number は #393/#548 由来の独立したデプロイ窓
-  // フォールバック（card-number-errors.ts、本Issue #834 のスコープ外）で、
-  // insertData から列を削るだけで RETURNING 自体は無指定のまま card_number を
+  // card_number は #393/#548 由来の独立したデプロイ窓フォールバック
+  // （card-number-errors.ts、本Issue #834 のスコープ外）で、insertData から
+  // 列を削るだけで RETURNING 自体は無指定のまま card_number を
   // 要求し続ける。#834 撤去前は末尾の CARDS_SAFE_COLUMNS（card_number も除外）が
   // これを最終的に救済していたが、card_number は本番・preview 実測済み（#834）
   // のため通常はこの分岐へ到達しない。真に card_number が欠落する環境では
   // このフォールバックは単独では完結しない点に注意（card-number-errors.ts 側の
-  // 対応が必要。本Issueでは意図的に手を加えない）。
+  // 対応が必要。本Issueでは意図的に手を加えない。要否の実測・整理は #954 参照）。
   if (error && isMissingCardNumberColumnError(error)) {
     delete insertData.card_number;
     ({ card, error } = await attemptInsert());
@@ -586,7 +586,7 @@ function resolveSortColumn(sortField: SortField): AnyColumn {
  *
  * #834: 「本番未デプロイ8列」（card_number/hp/atk/...）に対する SELECT 列
  * リストのフォールバックは、本番実測で8列とも実在することを確認したため撤去
- * した（cards-safe-columns.ts 参照）。
+ * した（card-padding-color-errors.ts 参照）。
  */
 async function fetchCardsFromDBPg(
   streamerId: string,
@@ -649,9 +649,9 @@ async function fetchCardsFromDBPg(
     rowsError = error;
   }
 
-  // Claude Auto Review 指摘: card_number ソート列のこのフォールバックも
-  // insertCardPg 側と同じ理由で、card_number が真に欠落する環境では単独では
-  // 完結しない（#834 のスコープ外。card-number-errors.ts 側の対応が必要）。
+  // card_number ソート列のこのフォールバックも insertCardPg 側と同じ理由で、
+  // card_number が真に欠落する環境では単独では完結しない（#834 のスコープ外。
+  // card-number-errors.ts 側の対応が必要。要否の実測・整理は #954 参照）。
   // card_number は本番・preview 実測済み（#834）のため通常はここへ到達しない。
   if (
     rowsError &&
