+++ tests/unit/api/cards-id-route-driver-parity.test.ts
@@ -18,7 +18,7 @@ import { getDb } from "@/lib/db/client";
 import { removeBlobFile } from "@/lib/storage-db";
 import { deleteFromR2 } from "@/lib/r2-client";
 import { sha256Prefix } from "@/lib/crypto-utils";
-import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR } from "@/lib/db/cards-safe-columns";
+import { CARDS_COLUMNS_WITHOUT_PADDING_COLOR } from "@/lib/db/card-padding-color-errors";
 
 // #830: 画像削除経路で参照される R2 の公開URL・削除操作を差し替える
 const R2_PUBLIC_URL = "https://images.example.test";
